requirements.txt
PyYAML
requests
pip3
pytest
allure-pytest
